[Refactor] 일부 오류 수정
